requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.18
chromadb==0.5.0
langchain==0.2.6
langchain-openai==0.1.14
langgraph==0.1.5
openai==1.35.0
fpdf2==2.7.9
qrcode==7.4.2
Pillow==10.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pyngrok==8.0.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
pytest==8.2.2
pytest-asyncio==0.23.7
bcrypt==4.1.3
supabase==2.4.3
psycopg2-binary==2.9.9
